move mapview to its own file and expose through index
